refactor: move playerCount logic out of useCollectionFilters
