+++ src/types.d.ts
@@ -1,6 +1,6 @@
 export type {
   CollectionFilterState,
   CollectionFilterReducer,
-} from "./components/SearchFilterForm/useCollectionFilters";
+} from "./components/SearchFilterForm/hooks/useCollectionFilters";
 
 export type { BoardGame } from "./components/SearchFilterForm/BggCollection/BggCollection.utils";
